publishing-service/requirements.txt
uvicorn==0.24.0.post1
fastapi==0.104.1
requests~=2.31.0
botocore~=1.32.6
pandas~=1.3.3
geopandas~=0.10.2
matplotlib~=3.7.4
jsonpickle~=2.0.0
pydantic~=2.5.2
python-multipart==0.0.6
PyJWT~=2.8.0
pytest~=6.2.5
h11~=0.14.0
pip~=23.3.1
toml~=0.10.2
wheel~=0.38.4
cryptography~=38.0.4
py~=1.11.0
lxml~=4.9.3
tornado~=6.3.3
anyio~=3.7.1
sniffio~=1.3.0
starlette~=0.27.0
click~=8.1.7
mypy~=0.910
decorator~=5.1.1
PyYAML~=6.0
itsdangerous~=2.1.2
Jinja2~=3.1.2
setuptools~=67.0.0
future~=0.18.3
boto3~=1.29.6
awscli~=1.30.6
pydantic-settings~=2.1.0
